BloxyBot (ISRP)/requirements.txt
aiodns==3.1.1
aiohttp==3.9.0
aiosignal==1.3.1
aiosqlite==0.20.0
anyio==3.7.1
asgiref==3.8.1
astunparse==1.6.3
asyncio==3.4.3
attrs==23.1.0
beautifulsoup4==4.12.3
blinker==1.7.0
braceexpand==0.1.7
Brotli==1.1.0
bs4==0.0.2
cachetools==5.3.3
cairocffi==1.6.1
CairoSVG==2.7.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cogwatch==3.3.1
colorama==0.4.6
cookies_discord_components==3.0.0
cryptography==41.0.7
cssselect2==0.7.0
defusedxml==0.7.1
discord==2.3.2
discord-py-interactions==5.11.0
discord-py-slash-command==1.1.2
discord-typings==0.7.0
discord.py==2.3.2
disnake==2.9.1
dnspython==2.4.2
emoji==2.8.0
fernet==1.0.1
frozenlist==1.4.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.4
import-expression==1.1.4
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.3
jishaku==2.5.2
lottie==0.7.0
markdown-it-py==3.0.0
markination==0.2.3
MarkupSafe==2.1.5
mdurl==0.1.2
motor==3.3.2
multidict==6.0.4
natural==0.2.0
numpy==1.26.4
oauthlib==3.2.2
orjson==3.9.10
outcome==1.3.0.post0
packaging==23.2
parsedatetime==2.6
Pillow==10.1.0
playwright==1.41.0
psutil==5.9.8
pyaes==1.6.1
pycares==4.4.0
pycparser==2.21
pyee==11.0.1
Pygments==2.17.2
PyJWT==2.8.0
pymongo==4.6.0
PyNaCl==1.5.0
PySocks==1.7.1
pystyle==2.9
python-dateutil==2.8.2
python-dotenv==1.0.0
random-username==1.0.2
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.0
roblox==2.0.0
schedule==1.2.1
selenium==4.5.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
tinycss2==1.2.1
tomli==2.0.1
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.1.0
watchdog==4.0.0
watchfiles==0.15.0
watchgod==0.8.2
webencodings==0.5.1
Werkzeug==3.0.2
wheel==0.42.0
wsproto==1.2.0
yarl==1.9.3